Refactor: Projects page
