+++ src/App.tsx
@@ -2,8 +2,8 @@ import { ApolloProvider } from '@apollo/client';
 import {createBrowserRouter, RouterProvider} from 'react-router-dom';
 import MenuContext from './contexts/MobileMenuContext';
 import { client } from './libs/apollo';
-import { Home } from './pages/Home';
-import Projects from './pages/Projects';
+import { Home, Projects } from './pages/';
+
 
 function App() {
   const router = createBrowserRouter([
