+++ src/pages/index.tsx
@@ -1,0 +1,2 @@
+export * from './Home';
+export * from './Projects';
